encryption/requirements.txt
# for encryption
pycryptodome
pyope

# for generation
numpy
